+++ src/main/java/net/geforcemods/securitycraft/compat/hudmods/WTHITClientDataProvider.java
@@ -1,8 +1,11 @@
 package net.geforcemods.securitycraft.compat.hudmods;
 
+import java.awt.Rectangle;
+
 import mcp.mobius.waila.api.IBlockAccessor;
 import mcp.mobius.waila.api.IBlockComponentProvider;
 import mcp.mobius.waila.api.IClientRegistrar;
+import mcp.mobius.waila.api.ICommonAccessor;
 import mcp.mobius.waila.api.IEntityAccessor;
 import mcp.mobius.waila.api.IEntityComponentProvider;
 import mcp.mobius.waila.api.IEventListener;
@@ -12,11 +15,13 @@ import mcp.mobius.waila.api.ITooltipComponent;
 import mcp.mobius.waila.api.IWailaClientPlugin;
 import mcp.mobius.waila.api.WailaConstants;
 import mcp.mobius.waila.api.component.ItemComponent;
+import net.geforcemods.securitycraft.ClientHandler;
 import net.geforcemods.securitycraft.api.IOwnable;
 import net.geforcemods.securitycraft.compat.IOverlayDisplay;
 import net.geforcemods.securitycraft.entity.AbstractSecuritySeaBoat;
 import net.geforcemods.securitycraft.entity.sentry.Sentry;
 import net.geforcemods.securitycraft.util.Utils;
+import net.minecraft.client.gui.GuiGraphicsExtractor;
 import net.minecraft.network.chat.Component;
 import net.minecraft.network.chat.MutableComponent;
 import net.minecraft.world.item.ItemStack;
@@ -70,10 +75,9 @@ public final class WTHITClientDataProvider extends HudModHandler implements IWai
 		addEntityInfo(data.getEntity(), data.getPlayer(), tooltip::addLine, config::getBoolean);
 	}
 
-	// TODO: Broken in WTHIT
-//	@Override
-//	public void onBeforeTooltipRender(GuiGraphics guiGraphics, Rectangle rectangle, ICommonAccessor accessor, IPluginConfig config, Canceller canceller) {
-//		if (ClientHandler.isPlayerMountedOnCamera())
-//			canceller.cancel();
-//	}
+	@Override
+	public void onBeforeTooltipRender(GuiGraphicsExtractor guiGraphics, Rectangle rectangle, ICommonAccessor accessor, IPluginConfig config, Canceller canceller) {
+		if (ClientHandler.isPlayerMountedOnCamera())
+			canceller.cancel();
+	}
 }