+++ Changelog.md
@@ -2,4 +2,4 @@
 
 - Fix: Screens that contain item slots do not darken the world in the background
 - Fix: SecurityCraft's custom JEI recipe categories do not show up correctly
-- Fix: The Jade overlay shows while viewing a Security Camera
+- Fix: The Jade and WTHIT overlays show while viewing a Security Camera